Apply review suggestions for SE-0162 test organization
  - Consolidate SE-0162 tests into PackageInfoMapperTests
  - Remove redundant PackageInfoMapperCustomLayoutTests file
  - Eliminate duplicate test coverage already handled by existing tests
